factor out team lookup
